% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages_to_user.R
\name{generate_quality_message}
\alias{generate_quality_message}
\title{Generate a quality message for the diagnostic table caption.}
\usage{
generate_quality_message(
  quality = "purple",
  sample_type = "cs",
  is_post_repair_valid = TRUE,
  post_repair_score = NULL
)
}
\arguments{
\item{quality}{A character string indicating the quality level ("purple", "green", "yellow", "red").}

\item{sample_type}{A character string, either "cs" or "eqam".}

\item{is_post_repair_valid}{A logical indicating if the data is valid for analysis \emph{after} repair attempts.}

\item{post_repair_score}{An integer, the new quality score after a successful repair.}
}
\value{
A single, randomly selected character string message.
}
\description{
Generate a quality message for the diagnostic table caption.
}
