% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_app_documentation.R
\name{get_app_documentation}
\alias{get_app_documentation}
\title{Application Documentation UI}
\usage{
get_app_documentation()
}
\description{
Generates the help documentation using glassHelp components.
}
