% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassLayout.R
\name{glassCol}
\alias{glassCol}
\title{Glass Grid Column}
\usage{
glassCol(width, ...)
}
\arguments{
\item{width}{Integer between 1 and 12. 12 is full width, 6 is half width.}

\item{...}{Content.}
}
\description{
A column within a \code{glassRow}.
}
