% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassButton.R
\name{glassDownloadButton}
\alias{glassDownloadButton}
\title{Glass Download Button}
\usage{
glassDownloadButton(
  outputId,
  label = "Download",
  icon = shiny::icon("download"),
  width = NULL,
  disabled = FALSE
)
}
\arguments{
\item{outputId}{The output variable to read the download from.}

\item{label}{The label to display.}

\item{icon}{An optional icon() to appear in the white section.}

\item{width}{The width of the button (e.g., '100\%', '200px').}

\item{disabled}{Logical. Should the button be disabled on load?}
}
\description{
A custom download button with the "Split Capsule" Glass theme.
}
