% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassLayout.R
\name{glassNavItem}
\alias{glassNavItem}
\title{Glass Navigation Item (Icon Button)}
\usage{
glassNavItem(tabName, icon, title, active = FALSE)
}
\arguments{
\item{tabName}{The ID of the page this button opens.}

\item{icon}{The icon to display.}

\item{title}{The tooltip text.}

\item{active}{Logical. Set TRUE for the default start page.}
}
\description{
Glass Navigation Item (Icon Button)
}
