% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassLayout.R
\name{glassPage}
\alias{glassPage}
\title{Glass Layout Shell}
\usage{
glassPage(title, branding, sidebar, header_items = NULL, ...)
}
\arguments{
\item{title}{App Title.}

\item{branding}{UI element for top-left.}

\item{sidebar}{UI element from \code{glassSidebar}.}

\item{header_items}{UI elements for header.}

\item{...}{\code{glassRouter} content.}
}
\description{
Glass Layout Shell
}
