% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassRadioButtons.R
\name{glassRadioButtons}
\alias{glassRadioButtons}
\title{Glass Radio Group Input}
\usage{
glassRadioButtons(
  inputId,
  label = NULL,
  label_icon = NULL,
  help_text = NULL,
  choices,
  selected = NULL,
  inline = TRUE,
  width = "100\%",
  disabled = FALSE
)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{label}{Optional display label text.}

\item{label_icon}{Optional \code{icon()} for the label.}

\item{help_text}{Optional tooltip text to display on hover.}

\item{choices}{List of values. Can be named vector c("Label" = "val").}

\item{selected}{The initially selected value.}

\item{inline}{(Deprecated) Always inline/responsive by design.}

\item{width}{The width of the container.}

\item{disabled}{Logical. Should it be disabled on load?}
}
\description{
A custom, responsive set of toggle buttons.
}
