% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassLayout.R
\name{glassRow}
\alias{glassRow}
\title{Glass Grid Row}
\usage{
glassRow(...)
}
\arguments{
\item{...}{Content (usually \code{glassCol} elements).}
}
\description{
A CSS Grid container that replaces \code{fluidRow}.
Automatically handles spacing (gap) between columns.
}
