% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassSlider.R
\name{glassSlider}
\alias{glassSlider}
\title{Glass Slider Input}
\usage{
glassSlider(
  inputId,
  label = NULL,
  choices,
  selected = NULL,
  unit = "",
  width = "100\%"
)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{label}{Optional label for the slider input.}

\item{choices}{Vector of values (numbers or strings).}

\item{selected}{The initially selected value.}

\item{unit}{Optional string to append to values (e.g. " dpi").}

\item{width}{The width of the input, e.g., '100\%'.}
}
\description{
Glass Slider Input
}
