% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassTabsetPanel.R
\name{glassTabsetPanel}
\alias{glassTabsetPanel}
\title{Glass Tabset Panel}
\usage{
glassTabsetPanel(inputId, ..., selected = NULL, color = "purple")
}
\arguments{
\item{inputId}{The id of the input object.}

\item{...}{One or more glassTabPanel() items.}

\item{selected}{The value of the tab to select initially.}

\item{color}{The active color theme: "purple" (default) or "green".}
}
\description{
Glass Tabset Panel
}
