% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassLoader.R
\name{hideGlassLoader}
\alias{hideGlassLoader}
\title{Hide Glass Loader}
\usage{
hideGlassLoader(
  id = "global_loader",
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{ID for loader-meldingen som skal skjules}

\item{session}{Shiny session object}
}
\description{
Hide Glass Loader
}
