% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassChart.R
\name{observeGlassChart}
\alias{observeGlassChart}
\title{Observe Glass Chart (Server)}
\usage{
observeGlassChart(outputId, expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{The ID used in glassChartOutput.}

\item{expr}{Expression returning list(cs=..., pb=..., ce=..., params=...)}
}
\description{
Call this directly in the server function (do NOT assign to output$id).
}
