% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks_and_repairs.R
\name{post_repair_diagnostics}
\alias{post_repair_diagnostics}
\title{Perform a Data Check After Repair}
\usage{
post_repair_diagnostics(
  data,
  prior_repair_diagnostics,
  methods_to_remove = NULL,
  type = "cs"
)
}
\arguments{
\item{data}{A data.table containing the raw clinical sample or EQAM data.}

\item{prior_repair_diagnostics}{Output from \code{commutability::check_data()}
based on the non-repaired data.}

\item{methods_to_remove}{A \code{character} vector or \code{NULL}. The methods
that are deemed invalid in the clinical sample data, the EQAM data or both.}

\item{type}{A \code{character} string. Either \code{'cs'} or \code{'eqam'}}
}
\value{
The output from \code{commutability::check_data(repaired_data)} or
\code{prior_repair_diagnostics}, if \code{commutability::repair_data(data)}
fails.
}
\description{
Perform a Data Check After Repair
}
