% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassFileInput.R
\name{processGlassFile}
\alias{processGlassFile}
\title{Helper to process Glass File in R}
\usage{
processGlassFile(input_data)
}
\arguments{
\item{input_data}{The value from input$yourId}
}
\description{
Since we aren't using Shiny's native uploader, the input returns a list
containing the Base64 content. This helper saves it to a temp file
so it behaves exactly like a standard fileInput.
}
