% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{render_diagnostic_table}
\alias{render_diagnostic_table}
\title{Render a styled diagnostic table using GlassTable}
\usage{
render_diagnostic_table(
  diagnostics,
  type = "cs",
  is_post_repair_valid = TRUE,
  post_repair_score = NULL
)
}
\arguments{
\item{diagnostics}{The diagnostic list object from commutability::check_data()}

\item{type}{A character string, either "cs" for clinical sample or "eq" for EQA material}

\item{is_post_repair_valid}{Logical}

\item{post_repair_score}{Integer}
}
\value{
A GlassTable UI object
}
\description{
Render a styled diagnostic table using GlassTable
}
