% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{select_best_model}
\alias{select_best_model}
\title{Step 2: Select Best Model based on Thresholds (Fast Operation)}
\usage{
select_best_model(
  candidate_results,
  zeta_critical = 4.24,
  parsimony_threshold = 0.2
)
}
\description{
Step 2: Select Best Model based on Thresholds (Fast Operation)
}
