% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassLoader.R
\name{showGlassLoader}
\alias{showGlassLoader}
\title{Show Glass Loader}
\usage{
showGlassLoader(
  id = "global_loader",
  text = NULL,
  selector = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{ID for loader-meldingen}

\item{text}{Valgfri tekst å vise under spinneren}

\item{selector}{(Valgfri) CSS-selector for elementet loaderen skal dekke (f.eks. "#mitt-kort-id").
Hvis NULL, dekker den hele skjermen.}

\item{session}{Shiny session object}
}
\description{
Show Glass Loader
}
