% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassChart.R
\name{updateGlassChart}
\alias{updateGlassChart}
\title{Update Glass Chart (Low-level)}
\usage{
updateGlassChart(session, inputId, cs_data, pb_data, ce_data, params)
}
\arguments{
\item{session}{The Shiny session object.}

\item{inputId}{The element ID (namespaced).}

\item{cs_data}{Clinical sample data frame.}

\item{pb_data}{Prediction band data frame.}

\item{ce_data}{Commutability evaluation data frame.}

\item{params}{List of visual parameters (colors, etc).}
}
\description{
Sends the data payload to the JavaScript via custom message.
Call this when you want to manually update the chart (e.g. inside an observer).
}
