% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassLoader.R
\name{useGlassLoader}
\alias{useGlassLoader}
\title{Glass Loader UI Component}
\usage{
useGlassLoader()
}
\value{
HTML tags to be inserted into the UI.
}
\description{
Adds a global, glass-morphism styled loading overlay to the application.
This overlay automatically appears when Shiny is busy (with a slight delay to prevent flickering).
}
