% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassButton.R
\name{glassButton}
\alias{glassButton}
\title{Glass Action Button (Split Style)}
\usage{
glassButton(inputId, label, icon = NULL, width = NULL, disabled = FALSE)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{label}{The label to display.}

\item{icon}{An optional icon() to appear in the white section.}

\item{width}{The width of the button (e.g., '100\%', '200px').}

\item{disabled}{Logical. Should the button be disabled on load?}
}
\description{
A custom, two-tone action button.
Left: White pill with Icon. Right: Purple gradient with Label.
}
