% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{render_diagnostic_table}
\alias{render_diagnostic_table}
\title{Render a styled diagnostic table}
\usage{
render_diagnostic_table(
  diagnostics,
  type = "cs",
  is_post_repair_valid = TRUE,
  post_repair_score = NULL
)
}
\arguments{
\item{diagnostics}{The diagnostic list object from commutability::check_data()}

\item{type}{A character string, either "cs" for clinical sample or "eq" for EQA material}
}
\value{
A kableExtra HTML table widget
}
\description{
Render a styled diagnostic table
}
