% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{render_diagnostic_text}
\alias{render_diagnostic_text}
\title{Render diagnostic test results with Glass styling}
\usage{
render_diagnostic_text(diagnostics, other_diagnostics, type = "cs")
}
\arguments{
\item{diagnostics}{The diagnostic list object}

\item{other_diagnostics}{The diagnostic list object from the other data type}

\item{type}{A character string, either "cs" or "eq"}
}
\description{
Render diagnostic test results with Glass styling
}
