% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{render_diagnostic_text}
\alias{render_diagnostic_text}
\title{Render diagnostic test results as styled HTML text}
\usage{
render_diagnostic_text(diagnostics, type = "cs")
}
\arguments{
\item{diagnostics}{The diagnostic list object from commutability::check_data()}

\item{type}{A character string, either "cs" for clinical sample or "eq" for EQA material}
}
\value{
An HTML object
}
\description{
Render diagnostic test results as styled HTML text
}
