% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassLoader.R
\name{useGlassLoader}
\alias{useGlassLoader}
\title{Glass Loader}
\usage{
useGlassLoader()
}
\arguments{
\item{id}{Unik ID for loaderen (namespace blir håndtert automatisk hvis session sendes)}
}
\value{
HTML for loader dependencies
}
\description{
Glass Loader
}
